POC-api/POC-api/requirements.txt
DateTime==4.3
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.1.0
zope.interface==4.7.1
django-heroku==0.3.1